requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-anthropic>=0.1.1
langchain-community>=0.0.13
openai>=1.3.0
anthropic>=0.8.0
faiss-cpu>=1.7.4
pypdf>=3.17.0
python-dotenv>=1.0.0
tiktoken>=0.5.1